添加查询博客/调整标签删除

+++ springboot-blog/src/main/java/com/yccztt/controller/label/LabelController.java
@@ -1,6 +1,8 @@
 package com.yccztt.controller.label;
 
+import com.yccztt.domain.article.Article;
 import com.yccztt.domain.label.Label;
+import com.yccztt.serivce.article.ArticleService;
 import com.yccztt.serivce.label.LabelService;
 import com.yccztt.utils.ResultUtil;
 import com.yccztt.utils.UuidUtil;
@@ -25,12 +27,16 @@ public class LabelController {
     @Autowired
     LabelService labelService;
 
+    @Autowired
+    ArticleService articleService;
+
     /**
      * 查询所有标签
      * @return
      */
     @GetMapping("/list")
     public ResultUtil queryLabel() {
+        System.out.println("查询所有标签====> queryLabel");
         List<Label> labelList = labelService.queryLabel();
         return ResultUtil.success(labelList);
     }
@@ -42,6 +48,7 @@ public class LabelController {
      */
     @GetMapping("/add/{label}")
     public ResultUtil addLabel(@PathVariable("label") String label) {
+        System.out.println("新增标签====> addLabel："+label);
         Label labelByName = labelService.queryLabelByName(label);
         if (labelByName==null) {
             SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
@@ -54,9 +61,22 @@ public class LabelController {
         }
     }
 
-    @GetMapping("/delete/{labelId}")
-    public ResultUtil deleteLabel(@PathVariable("labelId") String labelId) {
-        labelService.deleteLabel(labelId);
-        return ResultUtil.success();
+    /**
+     * 删除标签
+     * @param labelId
+     * @param labelName
+     * @return
+     */
+    @GetMapping("/delete/{labelId}/{labelName}")
+    public ResultUtil deleteLabel(@PathVariable("labelId") String labelId,
+                                  @PathVariable("labelName") String labelName) {
+        System.out.println("删除标签====> deleteLabel："+labelId+"，"+labelName);
+        List<Article> articles = articleService.queryBlogsByLabel(labelName);
+        if (!articles.isEmpty()) {
+            return ResultUtil.error("改标签存在文章，无法删除");
+        } else {
+            labelService.deleteLabel(labelId);
+            return ResultUtil.success();
+        }
     }
 }
